requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
joblib==1.2.0
numpy==1.24.2
pi==0.1.2
psycopg2==2.9.6
python-decouple==3.8
pytz==2023.3
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
